requirements.txt
click==7.1.2
kafka-python==2.0.2
typer==0.3.2
